requirements.txt
# Temporary placeholder to satisfy package builds that open requirements.txt
# You can remove this file after the package installs successfully.

# (Leave empty unless specific deps are required)

python-dotenv
fastapi
uvicorn
pydantic
openai
openai-agents
fastapi
uvicorn[standard]
python-dotenv
requests
supabase
psycopg2-binary
celery[redis]
redis
PyMuPDF        # fitz
python-docx
pandas
tqdm
openai         # or your LLM provider client
tiktoken       # optionally for token counting
pinecone-client 